requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3
python-dotenv==1.0.0
Werkzeug==2.3.7
PyPDF2==3.0.1
python-dateutil==2.8.2
gunicorn==21.2.0
pytest==7.4.2
pytest-flask==1.2.0
Pillow>=9.5.0
pytesseract>=0.3.10
selenium>=4.15.0
webdriver-manager>=4.0.0
captcha>=0.4.0
opencv-python>=4.8.0
numpy>=1.21.0
# Note: html module is built-in to Python 3.9+, no need to install separately 